# static/swagger.yaml
openapi: 3.0.3
info:
  title: System obsługi praktyk – REST API
  description: |
    REST API dla systemu zarządzania praktykami studenckimi ANS Elbląg.

    ## Zasoby
    - **students** – studenci praktyk (CRUD)
    - **internships** – praktyki studenckie (CRUD + filtrowanie)
    - **documents** – dokumenty powiązane z praktykami (GET / POST / DELETE / weryfikacja)

    ## Statusy HTTP
    | Kod | Znaczenie |
    |-----|-----------|
    | 200 | OK – żądanie przetworzone poprawnie |
    | 201 | Created – zasób został utworzony |
    | 400 | Bad Request – brakujące / nieprawidłowe dane |
    | 404 | Not Found – zasób nie istnieje |
    | 405 | Method Not Allowed |
    | 409 | Conflict – naruszenie unikalności |
    | 500 | Internal Server Error |
  version: "1.0.0"
  contact:
    name: ANS Elbląg – Dział Praktyk
servers:
  - url: http://localhost:5000
    description: Serwer deweloperski
...
components:
  schemas:
    Student:
      type: object
      properties:
        id:           { type: integer, example: 1 }
        first_name:   { type: string,  example: "Anna" }
        last_name:    { type: string,  example: "Kowalska" }
        index_number: { type: string,  example: "12345" }
        email:        { type: string,  example: "anna.kowalska@student.ans.edu.pl" }
        user_id:      { type: integer, nullable: true, example: null }
        created_at:   { type: string,  format: date-time }
        updated_at:   { type: string,  format: date-time }
    StudentCreate:
      type: object
      required: [first_name, last_name, index_number, email]
      properties:
        first_name:   { type: string, example: "Anna" }
        last_name:    { type: string, example: "Kowalska" }
        index_number: { type: string, example: "12345" }
        email:        { type: string, example: "anna.kowalska@student.ans.edu.pl" }
        user_id:      { type: integer, nullable: true }
...
    Error:
      type: object
      properties:
        error:  { type: string, example: "Brakujące wymagane pola: first_name." }
        status: { type: integer, example: 400 }
    Message:
      type: object
      properties:
        message: { type: string, example: "Student ID 1 został usunięty." }

# ─────────────────────────────────────────────────────────────────────────────
# Ścieżki
# ─────────────────────────────────────────────────────────────────────────────
paths:
...
  /api/students:
    get:
      tags: [students]
      summary: Lista studentów
      description: Zwraca wszystkich studentów, opcjonalnie filtrowanych po nazwisku / e-mailu / numerze indeksu.
      parameters:
        - name: q
          in: query
          description: Wyszukiwanie częściowe (nazwisko, e-mail, numer indeksu)
          schema: { type: string }
      responses:
        "200":
          description: Lista studentów
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:    { type: integer }
                  students:
                    type: array
                    items: { $ref: "#/components/schemas/Student" }

    post:
      tags: [students]
      summary: Dodaj studenta
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/StudentCreate" }
            example:
              first_name: "Anna"
              last_name: "Kowalska"
              index_number: "12345"
              email: "anna.kowalska@student.ans.edu.pl"
      responses:
        "201":
          description: Student utworzony
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Student" }
        "400":
          description: Brakujące / nieprawidłowe dane
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "409":
          description: Student z tym indeksem lub e-mailem już istnieje
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/students/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: integer }

    get:
      tags: [students]
      summary: Pobierz studenta
      responses:
        "200":
          description: Dane studenta
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Student" }
        "404":
          description: Student nie istnieje
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }

    put:
      tags: [students]
      summary: Zaktualizuj studenta
      description: Częściowa aktualizacja – podaj tylko pola do zmiany.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/StudentCreate" }
            example:
              email: "nowy.email@student.ans.edu.pl"
      responses:
        "200":
          description: Zaktualizowane dane studenta
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Student" }
        "400": { description: Błąd walidacji, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "404": { description: Nie znaleziono, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }

    delete:
      tags: [students]
      summary: Usuń studenta
      description: Kaskadowo usuwa powiązane praktyki i dokumenty.
      responses:
        "200":
          description: Potwierdzenie usunięcia
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Message" }
        "404": { description: Nie znaleziono, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }

  # ── /api/internships ───────────────────────────────────────────────────────
...